pipelines/audio/requirements.txt
whisperx>=3.1.1
pyannote.audio>=3.1.1
PyYAML>=6.0.1
